Flattened models

+++ src/main/kotlin/com/flatts/township/interfaces/SupplyMarker.kt
@@ -2,6 +2,6 @@ package com.flatts.township.interfaces
 
 interface SupplyMarker {
     var supply: String
-    var quantity: Int
+    var quantity: Double
     var iconClass: String
 }
